Add sorting by date/name/priority/id in TODOLIST

+++ app.js
@@ -7,7 +7,11 @@ let names = {
     listContainer: '#list',
     container: '#container',
     alert:'.alert',
-    priorityInput: '#priority-select'
+    priorityInput: '#priority-select',
+    sortByName: '.sort-name',
+    sortByDate: '.sort-date',
+    sortByPriority: '.sort-priority',
+    restore: '.restore'
 }
 
 //Task class
@@ -17,7 +21,7 @@ class Task{
         this.deadline = deadline;
         this.id = id;
         this.priority = priority;
-        this.todoHTML = '<div class="task priority-%priority-number%" id="task-%id%"><div><input type="checkbox" name="" id="do-task"><p>%description%</p></div><div><p class="date-tag">%date%</p><input type="checkbox" name="" class="delete remove-task"></div></div>';
+        this.todoHTML = '<div class="task priority-%priority-number%" id="task-%id%"><div><input type="checkbox" name="" class="do-task"><p>%description%</p></div><div><p class="date-tag">%date%</p><input type="checkbox" name="" class="delete remove-task"></div></div>';
         this.archHTML = '<div class="task" id="task-%id%"><div><p>%description%</p></div><div><input type="checkbox" name="" class="delete remove-task"></div></div>';
         
         //add priority
@@ -27,6 +31,8 @@ class Task{
 //UI class
 class UI{
     static displayTasks(){
+        const elements = document.getElementsByClassName("task");
+        while (elements.length > 0) elements[0].remove();
         const tasks = StoreToDo.getTasks();
         tasks.forEach((task)=>UI.addTaskToList(task,'todo'));
         const tasks2 = StoreDone.getTasks();
@@ -59,12 +65,14 @@ class UI{
                 el.parentElement.remove();
             else
                 el.parentElement.parentElement.remove();
+            UI.showAlert('Task removed','info');
         }
-        else if (el.id=='do-task'){
+        else if (el.className=='do-task'){
             console.log(el.parentElement.parentElement)
             el.parentElement.parentElement.remove();
+            UI.showAlert('Task done','info');
         }
-        UI.showAlert('Task removed','info');
+        
     }
     static showAlert(message, className){
         const div = document.createElement('div');
@@ -85,7 +93,7 @@ class UI{
 class Store {
     static getTasks() {
         let tasks;
-        const name = this.prototype.constructor.name
+        const name = this.prototype.constructor.name;
         if(localStorage.getItem(name) === null){
             tasks = [];
         }
@@ -97,7 +105,7 @@ class Store {
     static getSpecificTask(domID){
         const id = domID.split('-')[1];
         const tasks = this.getTasks();
-        let out 
+        let out;
         tasks.forEach((task,index)=>{
             if(task.id == id){
                 out = task;
@@ -108,38 +116,44 @@ class Store {
     static addTask(task) {
         const tasks = this.getTasks();
         tasks.push(task);
-        const name = this.prototype.constructor.name
-        localStorage.setItem(name,JSON.stringify(tasks))
+        const name = this.prototype.constructor.name;
+        localStorage.setItem(name,JSON.stringify(tasks));
     }
 
     static removeTask(domID) {
         let tasks = this.getTasks();
         const id = domID.split('-')[1];
-        const name = this.prototype.constructor.name
-        //console.log(name)
-        //console.log(tasks[0], ' almost deleted')
+        const name = this.prototype.constructor.name;
         tasks.forEach((task, index)=>{
-            //console.log(task , ' in for')
             if(task.id == id){
                 tasks.splice(index,1);
-                //console.log(tasks, ' left')
-                
             }
         });
 
         localStorage.setItem(name, JSON.stringify(tasks));
     }
+    
 }
 class StoreToDo extends Store{
-    
+    static sort_by(sort, desc){
+        const tasks = this.getTasks();
+        //const name = this.prototype.constructor.name;
+        tasks.sort(function(a,b) {
+            if(a[sort]<b[sort]) return -1;
+            if(a[sort]>b[sort]) return 1;
+            return 0;
+        });
+        if (desc) tasks.reverse();
+        localStorage.setItem('StoreToDo', JSON.stringify(tasks));
+    }
 }
 class StoreDone extends Store{
     
 }
 
 //Event: display tasks
-document.addEventListener('DOMContentLoaded', UI.displayTasks);
-
+//document.addEventListener('DOMContentLoaded', UI.displayTasks);
+UI.displayTasks()
 //Event: add a task
 document.querySelector(names.confirmInput).addEventListener('click', e=>{
     //Prevent actual submit
@@ -184,7 +198,8 @@ document.querySelector(names.confirmInput).addEventListener('click', e=>{
 document.querySelector(names.todoList).addEventListener('click', (e)=>{
 
     //target doest have delete then 'do task' 
-    if( !(e.target.className.includes('delete'))){
+    if((e.target.className.includes('do-task'))){
+        console.log(e.target)
         //Get accurate task
         const taskID = e.target.parentNode.parentNode.id
         
@@ -202,7 +217,7 @@ document.querySelector(names.todoList).addEventListener('click', (e)=>{
         StoreToDo.removeTask(domID); 
 
     }
-    else{
+    else if(e.target.className.includes('delete')){
         //Remove task from UI
         UI.deleteTask(e.target);
         //Remove task from the Store
@@ -220,3 +235,35 @@ document.querySelector(names.archiveList).addEventListener('click', (e)=>{
         domID = e.target.parentNode.parentNode.id
         StoreDone.removeTask(domID);
 });
+//SORTING
+//Sort elements in to do list by name
+const sort_name_btn = document.querySelector(names.sortByName);
+const sort_date_btn = document.querySelector(names.sortByDate);
+const sort_priority_btn = document.querySelector(names.sortByPriority);
+const restore_btn = document.querySelector(names.restore);
+
+let desc = false;
+sort_name_btn.addEventListener('click', ()=>{
+    StoreToDo.sort_by('description', desc);
+    UI.displayTasks();
+    desc = !desc;
+});
+
+sort_date_btn.addEventListener('click', ()=>{
+    StoreToDo.sort_by('date', desc);
+    UI.displayTasks();
+    desc = !desc;
+});
+
+sort_priority_btn.addEventListener('click', ()=>{
+    StoreToDo.sort_by('priority', desc);
+    UI.displayTasks();
+    desc = !desc;
+});
+
+restore_btn.addEventListener('click', ()=>{
+    StoreToDo.sort_by('id', desc);
+    UI.displayTasks();
+    desc = !desc;
+});
+
